Use cast to larger type for string formatting to allow compilation on Solaris
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@38459 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormFilePrefsManager.m
@@ -321,7 +321,7 @@ objectValueForTableColumn: (NSTableColumn *)aTableColumn
 
   if([[aTableColumn identifier] isEqual: @"item"])
     {
-      obj = [NSString stringWithFormat: @"#%"PRIiPTR,rowIndex+1];
+      obj = [NSString stringWithFormat: @"#%ld",(long int)rowIndex+1];
     }
   else if([[aTableColumn identifier] isEqual: @"description"])
     {
